Fix an obscure crash.
A crash was observed when: the scroller was displaying a single
message, the displayed message had been replaced, and the replacement
had expired.  This is easy enough to reproduce by pausing the
scroller ('0') for a sufficiently long time.  The visible glyph's
successor had been removed from the circular queue of glyphs, and a
suitable subsequent glyph could not be located.

Work around this by not removing a glyph from the circular queue if
its the ultimate successor of a visible glyph.  It's easy enough to
see if the glyph is the ultimate successor, but its harder to
determine if it has a visible predecessor, so give each glyph an
explicit predecessor link.  Also, rename visible_count to holder_count
to more accurately identify what it represents.